// src/types/enums/conditionsTypesEnum.ts
export enum conditionsTypes {
    PERCENT = 'percent',
    VIEWING = 'viewing',
    DAYS_AFTER = 'days_after',
    MINUTES_ON_SCREEN = 'viewing',
    SECTION_VALIDATED = 'sectionValidated',
    MODULE_VALIDATED = 'moduleValidated',
    NO_CONDITION = 'noCondition'
}

export enum accessConditionsTypes {
    NO_CONDITION = 'noCondition',
    VALID_MODULE = 'validModule',
    ACCESS_DATE = 'date'
}